+++ sketch-animated-koralgif/src/androidTest/kotlin/com/github/panpf/sketch/animated/koralgif/test/decode/GifDrawableDecoderTest.kt
@@ -42,6 +42,7 @@ import com.github.panpf.sketch.test.utils.getDrawableOrThrow
 import com.github.panpf.sketch.test.utils.toRequestContext
 import com.github.panpf.sketch.transform.AnimatedTransformation
 import com.github.panpf.sketch.transform.PixelOpacity
+import com.github.panpf.sketch.util.Rect
 import com.github.panpf.sketch.util.Size
 import kotlinx.coroutines.test.runTest
 import org.junit.runner.RunWith
@@ -282,7 +283,7 @@ class GifDrawableDecoderTest {
     private data object TranslucentAnimatedTransformation : AnimatedTransformation {
         override val key: String = "TranslucentAnimatedTransformation"
 
-        override fun transform(canvas: Any): PixelOpacity {
+        override fun transform(canvas: Any, bounds: Rect): PixelOpacity {
             return PixelOpacity.TRANSLUCENT
         }
     }